Moved directory finding code to common parser code
