+++ iridauploader/parsers/common.py
@@ -1,4 +1,4 @@
-from os import path
+import os
 from csv import reader
 
 from iridauploader.parsers import exceptions
@@ -19,7 +19,7 @@ def get_csv_reader(sample_sheet_file):
     returns a csv.reader object
     """
 
-    if path.isfile(sample_sheet_file):
+    if os.path.isfile(sample_sheet_file):
         csv_file = open(sample_sheet_file, "r")
         # strip any trailing newline characters from the end of the line
         # including Windows newline characters (\r\n)
@@ -33,3 +33,24 @@ def get_csv_reader(sample_sheet_file):
                                           sample_sheet_file)
 
     return csv_reader
+
+def find_directory_list(directory):
+        """Find and return all directories in the specified directory.
+
+        Arguments:
+        directory -- the directory to find directories in
+
+        Returns: a list of directories including current directory
+        """
+
+        # Checks if we can access to the given directory, return empty and log a warning if we cannot.
+        if not os.access(directory, os.W_OK):
+            raise exceptions.DirectoryError("The directory is not writeable, "
+                                            "can not upload samples from this directory {}".format(directory),
+                                            directory)
+
+        dir_list = next(os.walk(directory))[1]  # Gets the list of directories in the directory
+        full_dir_list = []
+        for d in dir_list:
+            full_dir_list.append(os.path.join(directory, d))
+        return full_dir_list
