+++ iridauploader/tests/parsers/miseq/test_parser.py
@@ -195,37 +195,3 @@ class TestGetSequencingRun(unittest.TestCase):
         res = Parser.get_sequencing_run(sample_sheet)
 
         self.assertEqual(type(res), model.SequencingRun)
-
-
-class TestFindDirectoryList(unittest.TestCase):
-    """
-    Test getting the list of directories
-    """
-
-    def setUp(self):
-        print("\nStarting " + self.__module__ + ": " + self._testMethodName)
-
-    def test_find_three(self):
-        """
-        Given a directory with 3 run directories in it, make sure all 3 directories are included in result
-        :return:
-        """
-        directory = path.join(path_to_module, "three_dirs")
-        dir_1 = path.join(directory, "first")
-        dir_2 = path.join(directory, "second")
-        dir_3 = path.join(directory, "third")
-        res = Parser._find_directory_list(directory)
-
-        self.assertIn(dir_1, res)
-        self.assertIn(dir_2, res)
-        self.assertIn(dir_3, res)
-
-    def test_find_none(self):
-        """
-        Given a directory with no sequencing run directories in it, make sure an empty list is returned
-        :return:
-        """
-        directory = path.join(path_to_module, "no_dirs")
-        res = Parser._find_directory_list(directory)
-
-        self.assertEqual(res, [])
